refactor ComponentWebotsRobot
The previous idea of create "Robots" for each device and threads to
update each "Robot" just to split components is not recommended. First
because it slow down the simulator; second because it goes against the
webots structure; third because it may produce unexpected behaviors.
One can now enable robot devices and tasks updates each one.

Data is available for other components. Component is not a device, but
an algorithm to use the raw data and other logics.